backend/requirements.txt
coreapi==2.3.1
coreschema==0.0.4
Django==1.11.5
django-after-response==0.2.2
django-bitfield==1.9.3
django-cors-headers==2.1.0
django-filter==1.0.4
django-jalali-date==0.2.6.1
django-redis-cache==1.7.1
django-redis-sessions==0.6
djangorestframework==3.7.7
djangorestframework-filters==0.10.1
googlemaps==2.5.1
gunicorn==19.7.1
lxml==3.8.0
mysqlclient==1.3.12
redis==2.10.6
requests==2.12.4
urllib3==1.22


